# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
...
        "/api/menu/:authID/getMenuByAuthID": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "AuthorityMenu"
                ],
                "summary": "获取用户动态路由",
                "parameters": [
                    {
                        "description": "空",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.Empty"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "获取用户动态路由,返回包括系统菜单详情列表",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/middleware.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/dto.SysBaseMenusResponse"
                                        },
                                        "msg": {
                                            "type": "string"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/api/menu/add_base_menu": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Menu"
                ],
                "summary": "新增菜单",
                "parameters": [
                    {
                        "description": "路由path, 父菜单ID, 路由name, 对应前端文件路径, 排序标记",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.SysBaseMenu"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "新增菜单",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/middleware.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "msg": {
                                            "type": "string"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/api/menu/add_menu_authority": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "AuthorityMenu"
                ],
                "summary": "增加menu和角色关联关系",
                "parameters": [
                    {
                        "description": "角色ID",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.AddMenuAuthorityInput"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "增加menu和角色关联关系",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/middleware.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "msg": {
                                            "type": "string"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/api/menu/getBaseMenuTree": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "AuthorityMenu"
                ],
                "summary": "获取用户动态路由",
                "parameters": [
                    {
                        "description": "空",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.Empty"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "获取系统菜单详情列表",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/middleware.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/dto.SysBaseMenusResponse"
                                        },
                                        "msg": {
                                            "type": "string"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/api/operation/delete_operations": {
            "delete": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "SysOperationRecord"
                ],
                "summary": "删除SysOperationRecord",
                "parameters": [
                    {
                        "description": "id值",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.IdsReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "删除SysOperationRecord",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/middleware.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "msg": {
                                            "type": "string"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/api/operation/get_operations": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "SysOperationRecord"
                ],
                "summary": "分页获取SysOperationRecord列表",
                "parameters": [
                    {
                        "type": "string",
                        "description": "关键字",
                        "name": "keyword",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "请求方法",
                        "name": "method",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "页码",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "每页大小",
                        "name": "pageSize",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "请求路径",
                        "name": "path",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "请求状态",
                        "name": "status",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "分页获取SysOperationRecord列表,返回包括列表,总数,页码,每页数量",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/middleware.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/dto.OperationListOutPut"
                                        },
                                        "msg": {
                                            "type": "string"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/api/operation/{id}/delete_operation": {
            "delete": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "SysOperationRecord"
                ],
                "summary": "删除SysOperationRecord",
                "parameters": [
                    {
                        "description": "空",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.Empty"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "删除SysOperationRecord",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/middleware.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "msg": {
                                            "type": "string"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "dto.AddMenuAuthorityInput": {
            "type": "object",
            "required": [
                "authorityId"
            ],
            "properties": {
                "authorityId": {
                    "description": "角色ID",
                    "type": "integer"
                },
                "menus": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.SysBaseMenu"
                    }
                }
            }
        },
...
        "dto.Empty": {
            "type": "object"
        },
        "dto.IdsReq": {
            "type": "object",
            "properties": {
                "ids": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                }
            }
        },
        "dto.OperationListOutPut": {
            "type": "object",
            "properties": {
                "keyword": {
                    "description": "关键字",
                    "type": "string"
                },
                "list": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.SysOperationRecord"
                    }
                },
                "page": {
                    "description": "页码",
                    "type": "integer"
                },
                "pageSize": {
                    "description": "每页大小",
                    "type": "integer"
                },
                "total": {
                    "type": "integer"
                }
            }
        },
...
        "dto.SysBaseMenusResponse": {
            "type": "object",
            "properties": {
                "menus": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.SysBaseMenu"
                    }
                }
            }
        },
...
        "middleware.Response": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {},
                "msg": {
                    "type": "string"
                },
                "real_err": {
                    "type": "string"
                }
            }
        },
...
        "model.SysAuthority": {
            "type": "object",
            "properties": {
                "authorityId": {
                    "description": "角色ID",
                    "type": "integer"
                },
                "authorityName": {
                    "description": "角色名",
                    "type": "string"
                },
                "children": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.SysAuthority"
                    }
                },
                "created_at": {
                    "type": "string"
                },
                "dataAuthorityId": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.SysAuthority"
                    }
                },
                "defaultRouter": {
                    "description": "默认菜单(默认dashboard)",
                    "type": "string"
                },
                "menus": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.SysBaseMenu"
                    }
                },
                "parentId": {
                    "description": "父角色ID",
                    "type": "integer"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
        "model.SysBaseMenu": {
            "type": "object",
            "properties": {
                "activeName": {
                    "type": "string"
                },
                "authoritys": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.SysAuthority"
                    }
                },
                "children": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.SysBaseMenu"
                    }
                },
                "closeTab": {
                    "description": "自动关闭tab",
                    "type": "boolean"
                },
                "created_at": {
                    "type": "string"
                },
                "disabled": {
                    "description": "是否在列表隐藏",
                    "type": "boolean"
                },
                "hidden": {
                    "description": "是否在列表隐藏",
                    "type": "boolean"
                },
                "icon": {
                    "description": "菜单图标",
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "keepAlive": {
                    "description": "是否缓存",
                    "type": "boolean"
                },
                "menuLevel": {
                    "type": "integer"
                },
                "name": {
                    "description": "路由name",
                    "type": "string"
                },
                "parentId": {
                    "description": "父菜单ID",
                    "type": "string"
                },
                "path": {
                    "description": "路由path",
                    "type": "string"
                },
                "sort": {
                    "description": "排序标记",
                    "type": "integer"
                },
                "title": {
                    "description": "菜单名",
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
        "model.SysOperationRecord": {
            "type": "object",
            "properties": {
                "agent": {
                    "description": "代理",
                    "type": "string"
                },
                "body": {
                    "description": "请求Body",
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "error_message": {
                    "description": "错误信息",
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "ip": {
                    "description": "请求ip",
                    "type": "string"
                },
                "latency": {
                    "description": "延迟",
                    "type": "string"
                },
                "method": {
                    "description": "请求方法",
                    "type": "string"
                },
                "path": {
                    "description": "请求路径",
                    "type": "string"
                },
                "resp": {
                    "description": "响应Body",
                    "type": "string"
                },
                "status": {
                    "description": "请求状态",
                    "type": "integer"
                },
                "updated_at": {
                    "type": "string"
                },
                "user": {
                    "$ref": "#/definitions/model.SysUser"
                },
                "user_id": {
                    "description": "用户id",
                    "type": "integer"
                }
            }
        },
        "model.SysUser": {
            "type": "object",
            "properties": {
                "DeptId": {
                    "type": "integer"
                },
                "activeColor": {
                    "description": "活跃颜色",
                    "type": "string"
                },
                "authorities": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.SysAuthority"
                    }
                },
                "authority": {
                    "$ref": "#/definitions/model.SysAuthority"
                },
                "authorityId": {
                    "description": "用户角色ID",
                    "type": "integer"
                },
                "avatar": {
                    "description": "用户头像",
                    "type": "string"
                },
                "baseColor": {
                    "description": "基础颜色",
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "email": {
                    "description": "用户邮箱",
                    "$ref": "#/definitions/sql.NullString"
                },
                "enable": {
                    "description": "用户是否被冻结 1正常 2冻结",
                    "type": "integer"
                },
                "id": {
                    "type": "integer"
                },
                "nickName": {
                    "description": "用户昵称",
                    "type": "string"
                },
                "phone": {
                    "description": "用户手机号",
                    "$ref": "#/definitions/sql.NullString"
                },
                "sideMode": {
                    "description": "用户侧边主题",
                    "type": "string"
                },
                "status": {
                    "$ref": "#/definitions/sql.NullInt64"
                },
                "updated_at": {
                    "type": "string"
                },
                "userName": {
                    "description": "用户登录名",
                    "type": "string"
                },
                "uuid": {
                    "description": "用户UUID",
                    "type": "string"
                }
            }
        },
        "sql.NullInt64": {
            "type": "object",
            "properties": {
                "int64": {
                    "type": "integer"
                },
                "valid": {
                    "description": "Valid is true if Int64 is not NULL",
                    "type": "boolean"
                }
            }
        },
        "sql.NullString": {
            "type": "object",
            "properties": {
                "string": {
                    "type": "string"
                },
                "valid": {
                    "description": "Valid is true if String is not NULL",
                    "type": "boolean"
                }
            }
        }
    }
}